remove route handler and call meerkat server directly, now that CORS is enabled
