+++ env.template
@@ -7,5 +7,5 @@ BETTER_AUTH_URL="http://localhost:3000"
 STAGE="local"
 
 # Meerkat API
-MEERKAT_USERNAME=""
-MEERKAT_PASSWORD=""
+NEXT_PUBLIC_MEERKAT_USERNAME=""
+NEXT_PUBLIC_MEERKAT_PASSWORD=""
