+++ package.json
@@ -63,7 +63,6 @@
     "tailwind-merge": "^3.3.1",
     "vaul": "^1.1.2",
     "zod": "^4.1.12",
-    "zukeeper": "^1.0.2",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
